refactor(name-resolve): Move FilteredWalker to ast_node.py
