+++ .idea/inspectionProfiles/project_inspections.xml
@@ -14,6 +14,7 @@
         </value>
       </option>
     </inspection_tool>
+    <inspection_tool class="PyInconsistentReturnsInspection" enabled="true" level="INFORMATION" enabled_by_default="true" editorAttributes="INFORMATION_ATTRIBUTES" />
     <inspection_tool class="PyPep8Inspection" enabled="true" level="WEAK WARNING" enabled_by_default="true">
       <option name="ignoredErrors">
         <list>
